third_party/blink/web_tests/platform/generic/dom/node/contains-method-expected.txt

This test checks behavior of Node.contains.

[test1 [test 2]]

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof document.contains is "function"
PASS test1.contains(test2) is true
PASS test1.contains(test1TextChild) is true
PASS document.contains(test1) is true
PASS document.contains(test1TextChild) is true
PASS document.contains(document) is true
PASS test1.contains(test1) is true
PASS test1TextChild.contains(test1TextChild) is true
PASS test1.contains(123) threw exception TypeError: Failed to execute 'contains' on 'Node': parameter 1 is not of type 'Node'..
PASS test1.contains(null) is false
PASS document.contains(iframe) is false
PASS iframe.contains(iframe) is true
iframe is now in document
PASS document.contains(iframe) is true
PASS document.contains(iframeDoc) is false
PASS iframeDoc.contains(iframeBody) is true
PASS document.contains(iframeBody) is false
PASS successfullyParsed is true

TEST COMPLETE
